Neural gas algorithm fixed.

+++ quantization/src/main/java/pl/iad/quantization/algorithms/gas/parameters/learning/PowerLearningFactor.java
@@ -10,7 +10,7 @@ public class PowerLearningFactor implements LearningFactorProvider {
     
     public final static double DEFAULT_MINIMUM_FACTOR = 0.005;
     
-    private final int maxEpochIndex;
+    private final double maxEpochIndex;
 
     private final double initialFactor;
